fix: routing issues to solution details
