+++ src/components/grid/Grid.js
@@ -1,5 +1,5 @@
 import { Card } from '../card/Card.js'
-import { getChallenges } from '../../services/challenges.js'
+import { getChallenges } from '@/services/challenges.js'
 
 export const Grid = async (app) => {
   const challenges = await getChallenges()
