+++ src/pages/solution_details/index.html
