+++ style.css
@@ -29,7 +29,7 @@ body {
 
 #app {
   width: 1440px;
-  max-width: 1440px;
+  max-width: 1024px;
   margin: 0 auto;
   padding: 2rem;
   text-align: center;
