+++ vite.config.js
@@ -1,0 +1,21 @@
+import { fileURLToPath, URL } from "url";
+import { defineConfig } from "vite";
+import { resolve } from "path";
+
+// https://vitejs.dev/config/
+export default defineConfig({
+  build: {
+    rollupOptions: {
+      input: {
+        main: resolve(__dirname, 'index.html'),
+        details: resolve(__dirname, 'src/pages/solution_details/index.html'),
+      },
+    },
+  },
+  resolve: {
+    alias: [
+      { find: '@', replacement: fileURLToPath(new URL('./src', import.meta.url)) },
+      { find: '@pages', replacement: fileURLToPath(new URL('./src/pages', import.meta.url)) }
+    ],
+  },
+});
